CodeBufferSpringBootMongoDB/src/test/java/org/dailycodebuffer/codebufferspringbootmongodb/baeldung/callback_test1/ServiceExample.java
package org.dailycodebuffer.codebufferspringbootmongodb.baeldung.callback_test1;

public class ServiceExample implements Service {

    @Override
    public void doAction(String request, Callback<Response> callback) {
        // Simulate some asynchronous action or network request
        // In a real-world scenario, you would perform an actual operation and invoke the callback based on the result.

        // For demonstration purposes, we'll create a simple response object
        Response response = new Response();
        response.setData("Sample data");

        // Simulate a successful response
        callback.onSuccess(response);

        // Alternatively, simulate a failure
         callback.onFailure("Request failed");
    }

    // Additional methods or logic as needed

    public void handleResponse(Response response) {
        // Process the response
        System.out.println("Handling response: " + response.getData());
    }
}
